feat:MNST-10 review test

+++ src/main/java/com/spring/boot/project/ms/manga/store/application/dto/response/MangaResponseDto.java
@@ -9,7 +9,7 @@ public record MangaResponseDto(
 
         String description,
 
-        Double totalVolumes,
+        Integer totalVolumes,
 
         LocalDate startDate
 ) {
